@@ -1,12 +1,12 @@
 module github.com/makiuchi-d/arelo
 
-go 1.18
+go 1.19
 
 require (
-	github.com/bmatcuk/doublestar/v4 v4.0.3
+	github.com/bmatcuk/doublestar/v4 v4.2.0
 	github.com/fsnotify/fsnotify v1.5.4
 	github.com/spf13/pflag v1.0.5
 	golang.org/x/xerrors v0.0.0-20220609144429-65e65417b02f
 )
 
-require golang.org/x/sys v0.0.0-20220702020025-31831981b65f // indirect
+require golang.org/x/sys v0.0.0-20220906135438-9e1f76180b77 // indirect
